+++ android/src/main/java/io/cobrowse/reactnative/CobrowseIOModule.java
@@ -28,12 +28,17 @@ import androidx.annotation.NonNull;
 import androidx.core.util.Predicate;
 import io.cobrowse.CobrowseIO;
 import io.cobrowse.Session;
+import io.cobrowse.CobrowseAccessibilityService;
 
-public class CobrowseIOModule extends ReactContextBaseJavaModule implements CobrowseIO.Delegate, CobrowseIO.SessionRequestDelegate, CobrowseIO.RedactionDelegate {
+public class CobrowseIOModule extends ReactContextBaseJavaModule
+  implements CobrowseIO.Delegate, CobrowseIO.SessionRequestDelegate,
+    CobrowseIO.SessionLoadDelegate, CobrowseIO.RedactionDelegate,
+    CobrowseIO.RemoteControlRequestDelegate {
 
-    private static final String SESSION_UPDATED = "session_updated";
-    private static final String SESSION_ENDED = "session_ended";
-    private static final String SESSION_REQUESTED = "session_requested";
+    private static final String SESSION_LOADED = "session.loaded";
+    private static final String SESSION_UPDATED = "session.updated";
+    private static final String SESSION_ENDED = "session.ended";
+    private static final String SESSION_REQUESTED = "session.requested";
 
     private final HashSet<Integer> unredactedTags = new HashSet<>();
     private NativeViewHierarchyManager nodeManager;
@@ -61,12 +66,10 @@ public class CobrowseIOModule extends ReactContextBaseJavaModule implements Cobr
     }
 
     @Override
-    public Map<String, Object> getConstants() {
-        final Map<String, Object> constants = new HashMap<>();
-        constants.put("SESSION_UPDATED", SESSION_UPDATED);
-        constants.put("SESSION_ENDED", SESSION_ENDED);
-        constants.put("SESSION_REQUESTED", SESSION_REQUESTED);
-        return constants;
+    public void sessionDidLoad(@NonNull Session session) {
+        getReactApplicationContext()
+                .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
+                .emit(SESSION_LOADED, Utility.convert(session));
     }
 
     @Override
@@ -112,6 +115,12 @@ public class CobrowseIOModule extends ReactContextBaseJavaModule implements Cobr
             }
             return unredacted;
         }
+    }
+
+    @Override
+    public void handleRemoteControlRequest(@NonNull Activity activity, @NonNull Session session) {
+      // no-op, this will be handed on the JS side via an "updated" event handler
+      // this stub just disables the default native prompt in the SDK
     }
 
     @Override
@@ -232,14 +241,14 @@ public class CobrowseIOModule extends ReactContextBaseJavaModule implements Cobr
     }
 
     @ReactMethod
-    public void loadSession(final String idOrCode, final Promise promise) {
+    public void getSession(final String idOrCode, final Promise promise) {
         Handler handler = new Handler(getReactApplicationContext().getMainLooper());
         handler.post(new Runnable() {
             public void run() {
                 CobrowseIO.instance().getSession(idOrCode, new io.cobrowse.Callback<Error, Session>() {
                     @Override
                     public void call(Error error, Session session) {
-                        if (error != null) promise.reject("cbio_load_session_failed", error);
+                        if (error != null) promise.reject("cbio_get_session_failed", error);
                         else promise.resolve(Conversion.convert(session));
                     }
                 });
@@ -289,4 +298,55 @@ public class CobrowseIOModule extends ReactContextBaseJavaModule implements Cobr
         });
     }
 
+    @ReactMethod
+    public void updateSession (final ReadableMap options, final Promise promise) {
+        Handler handler = new Handler(getReactApplicationContext().getMainLooper());
+        handler.post(new Runnable() {
+            public void run() {
+                Session current = CobrowseIO.instance().currentSession();
+
+                if (current == null) {
+                    promise.resolve(null);
+                    return;
+                }
+
+                if (options.hasKey("full_device")) {
+                    boolean fullDevice = options.getBoolean("full_device");
+
+                    current.setFullDevice(fullDevice, new io.cobrowse.Callback<Error, Session>() {
+                        @Override
+                        public void call(Error error, Session session) {
+                            if (error != null) promise.reject("cbio_full_device_failed", error);
+                            else promise.resolve(null);
+                        }
+                    });
+
+                    return;
+                }
+
+                if (options.hasKey("remote_control")) {
+                    String remoteControl = options.getString("remote_control");
+
+                    current.setRemoteControl(Utility.remoteControl(remoteControl), new io.cobrowse.Callback<Error, Session>() {
+                        @Override
+                        public void call(Error error, Session session) {
+                            if (error != null) promise.reject("cbio_remote_control_failed", error);
+                            else promise.resolve(null);
+                        }
+                    });
+                }
+            }
+        });
+    }
+
+    @ReactMethod
+    public void accessibilityServiceShowSetup (final Promise promise) {
+        CobrowseAccessibilityService.showSetup(getReactApplicationContext());
+        promise.resolve(null);
+    }
+
+    @ReactMethod
+    public void accessibilityServiceIsRunning(final Promise promise) {
+        promise.resolve(CobrowseAccessibilityService.isRunning(getReactApplicationContext()));
+    }
 }