+++ ios/CBIOSession+Bridging.m
@@ -1,4 +1,6 @@
 @import CobrowseIO;
+#import "CBIORCTUtil.h"
+
 
 @implementation CBIOSession (Bridging)
 
@@ -8,6 +10,7 @@
         @"code": self.code ? self.code : NSNull.null,
         @"state": self.state ? self.state : NSNull.null,
         @"full_device": @(self.fullDevice),
+        @"remote_control": [CBIORCTUtil remoteControl: self.remoteControl],
         @"agent": self.hasAgent ? @{
             @"name": self.agent.name,
             @"id": self.agent.id