+++ js/CobrowseIO.js
@@ -1,38 +1,67 @@
 import { Alert } from 'react-native'
+import Session from './Session'
 const CobrowseIONative = require('react-native').NativeModules.CobrowseIO
 const NativeEventEmitter = require('react-native').NativeEventEmitter
 
 const emitter = new NativeEventEmitter(CobrowseIONative)
 
 export default class CobrowseIO {
+  /** @deprecated */
   static get SESSION_UPDATED () {
-    return CobrowseIONative.SESSION_UPDATED
+    return 'session.updated'
   }
 
+  /** @deprecated */
   static get SESSION_ENDED () {
-    return CobrowseIONative.SESSION_ENDED
-  }
-
-  static get SESSION_REQUESTED () {
-    return CobrowseIONative.SESSION_REQUESTED
+    return 'session.ended'
   }
 
   static handleSessionRequest (session) {
+    if (this._sessionRequestShown) return
+    this._sessionRequestShown = true
     Alert.alert(
       'Support Request',
       'A support agent would like to use this app with you. Do you accept?',
       [{
         text: 'Reject',
-        onPress: () => this.endSession(),
+        onPress: () => {
+          this._sessionRequestShown = false
+          session.end()
+        },
         style: 'cancel'
       }, {
         text: 'Accept',
-        onPress: () => this.activateSession()
-      }], { cancelable: true })
+        onPress: () => {
+          this._sessionRequestShown = false
+          session.activate()
+        }
+      }], { cancelable: false })
+  }
+
+  static handleRemoteControlRequest (session) {
+    if (this._remoteControlRequestShown) return
+    this._remoteControlRequestShown = true
+    Alert.alert(
+      'Remote Control Request',
+      'A support agent would like to take remote control of this app. Do you accept?',
+      [{
+        text: 'Reject',
+        onPress: () => {
+          this._remoteControlRequestShown = false
+          session.setRemoteControl('rejected')
+        },
+        style: 'cancel'
+      }, {
+        text: 'Accept',
+        onPress: () => {
+          this._remoteControlRequestShown = false
+          session.setRemoteControl('on')
+        }
+      }], { cancelable: false })
   }
 
   static addListener (event, cb) {
-    return emitter.addListener(event, cb)
+    return emitter.addListener(event, (session) => cb(new Session(session)))
   }
 
   static start () {
@@ -60,26 +89,36 @@ export default class CobrowseIO {
   }
 
   static currentSession () {
-    return CobrowseIONative.currentSession()
+    return CobrowseIONative.currentSession().then((session) => session ? new Session(session) : null)
   }
 
   static createSession () {
-    return CobrowseIONative.createSession()
-  }
-
-  static activateSession () {
-    return CobrowseIONative.activateSession()
+    return CobrowseIONative.createSession().then((session) => session ? new Session(session) : null)
   }
 
   static getSession (codeOrId) {
-    return CobrowseIONative.getSession(codeOrId)
+    return CobrowseIONative.getSession(codeOrId).then((session) => session ? new Session(session) : null)
   }
 
+  /** @deprecated */
+  static activateSession () {
+    return CobrowseIONative.activateSession().then((session) => session ? new Session(session) : null)
+  }
+
+  /** @deprecated */
   static endSession () {
     return CobrowseIONative.endSession()
   }
 }
 
-CobrowseIO.addListener(CobrowseIO.SESSION_REQUESTED, (session) => {
+// the session.requested event is considered internal, it should
+// not be used outside these bindings
+CobrowseIO.addListener('session.requested', (session) => {
   CobrowseIO.handleSessionRequest(session)
+})
+
+CobrowseIO.addListener('session.updated', (session) => {
+  if (session.isActive() && session.remote_control === 'requested') {
+    CobrowseIO.handleRemoteControlRequest(session)
+  }
 })